Auth categories, user and answer routers

+++ src/controllers/answers.controller.ts
@@ -1,6 +1,7 @@
 import Database, { dataResponse, options, queryParam } from '../database';
 import dbConfig from '../db.config';
 import Errors from '../database/messages/errors';
+import JwtController from '../function/jws';
 
 interface answer {
   id?: number;
@@ -12,6 +13,7 @@ interface answer {
 
 const db = new Database('answers', dbConfig);
 const errors = new Errors();
+const jwtCtrl = new JwtController();
 
 const answerData = (Answer: answer): queryParam => [
     { selector: 'user_id', value: Answer.user_id },
@@ -25,20 +27,38 @@ class AnswersController {
         return await db.select('*', id ? options : null);
     }
 
-    async post(Answer: answer): Promise<dataResponse> {
+    async post(Answer: answer, auth?: string): Promise<dataResponse> {
         const { answer, parent_id, user_id } = Answer;
-        if (parent_id && user_id && answer) return db.insert(answerData(Answer));
-        else return errors.allNeeded;
+        if (auth && await jwtCtrl.checkToken(auth, {id:user_id})) {
+            if (parent_id && user_id && answer) 
+                return db.insert(answerData(Answer));
+            else return errors.allNeeded;
+        } else {
+            return errors.notAuth;
+        }
     }
 
-    async update(id: number, Answer: answer): Promise<dataResponse> {
+    async update(
+        id: number,
+        Answer: answer,
+        auth?: string
+    ): Promise<dataResponse> {
         const { answer, parent_id, user_id } = Answer;
-        if (answer || parent_id || user_id) return db.update(answerData(Answer), { selector: 'id', value: id });
-        else return errors.requestEmpty;
+        if (auth && await jwtCtrl.checkToken(auth, {id:user_id})) {
+            if (answer || parent_id || user_id)
+                return db.update(answerData(Answer), { selector: 'id', value: id });
+            else return errors.requestEmpty;
+        } else {
+            return errors.notAuth;
+        }
     }
 
-    async delete(id: number): Promise<dataResponse> {
-        return db.delete({ selector: 'id', value: id });
+    async delete(id: number, auth?: string): Promise<dataResponse> {
+        const check = await this.get(id);
+        const { user_id } = check.data;
+        if (auth && await jwtCtrl.checkToken(auth, {id:user_id}))
+            return db.delete({ selector: 'id', value: id });
+        else return errors.notAuth;
     }
 }
